fixed deriving->derive for nightly, cleaned up uint->usize

+++ examples/fromseed.rs
@@ -7,9 +7,9 @@ use noisy::gen::{ NoiseGen, Simplex };
 // Seed used to create a random number generator.
 static SEED: [u32; 4] = [9123678, 7890124, 6123462, 2789431];
 // Width of the output in columns.
-static WIDTH: uint = 80;
+static WIDTH: usize = 80;
 // Height of the output in rows.
-static HEIGHT: uint = 80;
+static HEIGHT: usize = 80;
 
 // A five color gradient used for the output.
 static GRADIENT: [&'static str; 5] = [" ", "░", "▒", "▓", "█"];
@@ -38,7 +38,7 @@ fn main() {
             val = val * 5.0;
 
             // Print the columns in the row.
-            print!("{}", GRADIENT[val as uint]);
+            print!("{}", GRADIENT[val as usize]);
         }
         // Start a new row.
         println!("");
